+++ pychron/processing/plotters/spectrum/spectrum.py
@@ -49,13 +49,19 @@ class Spectrum(BaseArArFigure):
 
         graph = self.graph
 
+        ag = self.analysis_group
         for pid, (plotobj, po) in enumerate(zip(graph.plots, plots)):
 
             plot = getattr(self, '_plot_{}'.format(po.plot_name))(po, plotobj, pid)
             if legend and po.plot_name == 'age_spectrum':
-                ident = self.analysis_group.identifier
-                sample = self.analysis_group.sample
+                ident = ag.identifier
+                sample = ag.sample
+
                 key = self.options.make_legend_key(ident, sample)
+                if key in legend.plots:
+                    ident = '{}-{:02n}'.format(ident, ag.aliquot)
+                    key = self.options.make_legend_key(ident, sample)
+
                 legend.plots[key] = plot
 
         self.graph.set_x_title('Cumulative %39ArK', plotid=0)
@@ -104,10 +110,10 @@ class Spectrum(BaseArArFigure):
         ag.plateau_age_error_kind = self.options.plateau_age_error_kind
         ag.pc_nsteps = self.options.pc_nsteps
         ag.pc_gas_fraction = self.options.pc_gas_fraction
-        if self.options.force_plateau:
-            ag.force_plateau_steps = self.options.force_plateau_start, self.options.force_plateau_end
+        if self.options.calculate_fixed_plateau:
+            ag.calculate_fixed_plateau_steps = self.options.calculate_fixed_plateau_start, self.options.calculate_fixed_plateau_end
         else:
-            ag.force_plateau_steps = tuple()
+            ag.calculate_fixed_plateau_steps = tuple()
 
         pma = None
         plateau_age = ag.plateau_age
@@ -231,10 +237,15 @@ class Spectrum(BaseArArFigure):
 
         a = self.options.envelope_alpha * 0.01
         sp = SpectrumErrorOverlay(component=ds,
-                                  spectrum=self,
+                                  use_user_color=self.options.user_envelope_color,
+                                  user_color=self.options.envelope_color,
                                   alpha=max(min(1.0, a), 0.0),
                                   use_fill=self.options.use_error_envelope_fill,
                                   nsigma=ns)
+
+        if self.options.user_envelope_color:
+            sp.user_color = self.options.envelope_color
+
         ds.underlays.append(sp)
         self.spectrum_overlays.append(sp)
 
@@ -242,13 +253,16 @@ class Spectrum(BaseArArFigure):
             lo = SpectrumLabelOverlay(component=ds,
                                       nsigma=ns,
                                       spectrum=self,
+                                      use_user_color=self.options.user_envelope_color,
+                                      user_color=self.options.envelope_color,
+
                                       font_size=self.options.step_label_font_size,
                                       display_extract_value=self.options.display_extract_value,
                                       display_step=self.options.display_step)
 
             ds.underlays.append(lo)
         # if po.scale == 'log':
-        #     p.value_axis.tick_generator = SparseLogTicks()
+        # p.value_axis.tick_generator = SparseLogTicks()
         # else:
         #     p.value_axis.tick_generator = SparseTicks()
         return ds
@@ -257,19 +271,20 @@ class Spectrum(BaseArArFigure):
     # overlays
     # ===============================================================================
     def _add_plateau_overlay(self, lp, bounds, plateau_age, ages, age_errors, info_txt):
+        opt = self.options
         ov = PlateauOverlay(component=lp, plateau_bounds=bounds,
                             cumulative39s=hstack(([0], self.xs)),
                             info_txt=info_txt,
                             id='plateau',
                             ages=ages,
                             age_errors=age_errors,
-                            line_width=self.options.plateau_line_width,
+                            line_width=opt.plateau_line_width,
 
-                            line_color=lp.color,
-                            # line_color=self.options.plateau_line_color,
-                            extend_end_caps=self.options.extend_plateau_end_caps,
-                            label_visible=self.options.display_plateau_info,
-                            label_font_size=self.options.plateau_font_size,
+                            line_color=opt.plateau_line_color if opt.user_plateau_line_color else lp.color,
+                            # line_color=opt.plateau_line_color,
+                            extend_end_caps=opt.extend_plateau_end_caps,
+                            label_visible=opt.display_plateau_info,
+                            label_font_size=opt.plateau_font_size,
 
                             # label_offset=plateau_age.std_dev*self.options.step_nsigma,
                             y=plateau_age.nominal_value * 1.25)
@@ -304,9 +319,9 @@ class Spectrum(BaseArArFigure):
             ai.temp_status = i in sel
 
         ag.dirty = True
-        tga = ag.integrated_age
-        mswd = ag.get_mswd_tuple()
-        text = self._build_integrated_age_label(tga, *mswd)
+        # tga = ag.integrated_age
+        # mswd = ag.get_mswd_tuple()
+        text = self._build_integrated_age_label(ag.integrated_age, ag.nanalyses)
         self.integrated_label.text = text
 
         if ag.plateau_age and self.plateau_overlay:
@@ -325,8 +340,8 @@ class Spectrum(BaseArArFigure):
         plateau_mswd, valid_mswd, nsteps = ag.get_plateau_mswd_tuple()
 
         e = plateau_age.std_dev * self.options.nsigma
-        text = self._build_label_text(plateau_age.nominal_value, e,
-                                      plateau_mswd, valid_mswd, nsteps,
+        text = self._build_label_text(plateau_age.nominal_value, e, nsteps,
+                                      mswd_args=(plateau_mswd, valid_mswd, nsteps),
                                       sig_figs=self.options.plateau_sig_figs)
 
         sample = self.analysis_group.sample
@@ -345,8 +360,9 @@ class Spectrum(BaseArArFigure):
     def _make_integrated_text(self):
         ag = self.analysis_group
         tga = ag.integrated_age
-        mswd = ag.get_mswd_tuple()
-        text = self._build_integrated_age_label(tga, *mswd)
+        n = ag.nanalyses
+        # mswd = ag.get_mswd_tuple()
+        text = self._build_integrated_age_label(tga, n)
         return text
 
     def _get_age_errors(self, ans):
@@ -420,14 +436,15 @@ class Spectrum(BaseArArFigure):
     # ===============================================================================
     # labels
     # ===============================================================================
-    def _build_integrated_age_label(self, tga, *args):
+    def _build_integrated_age_label(self, tga, n):
         txt = 'NaN'
         if not isnan(nominal_value(tga)):
             age, error = tga.nominal_value, tga.std_dev
-            error *= self.options.nsigma
-            txt = self._build_label_text(age, error, *args, sig_figs=self.options.integrated_sig_figs)
 
-        return 'Integrated Age= {}'.format(txt)
+            error *= self.options.nsigma
+            txt = self._build_label_text(age, error, n, sig_figs=self.options.integrated_sig_figs)
+
+        return u'Integrated Age= {}'.format(txt)
         # ============= EOF =============================================
         # def _get_plateau(self, analyses, exclude=None):
         # if exclude is None:
@@ -445,7 +462,7 @@ class Spectrum(BaseArArFigure):
         # if n > 1:
         # ans = []
         #
-        #     for j, ai in enumerate(analyses):
+        # for j, ai in enumerate(analyses):
         #         if j not in exclude and platbounds[0] <= j <= platbounds[1]:
         #             ans.append(ai)
         #             #            ans=[ai for (j,ai) in analyses if]